requirements.txt
faster-whisper==0.10.0
PyAudio==0.2.14
keyboard==0.13.5
requests==2.31.0
python-dotenv==1.0.1
edge-tts==6.1.9
langdetect==1.0.9
pygame==2.5.2 